device-generic-sdm845: new aports
Selecting this device in pmbootstrap should produce images which can be
flashed to any sdm845 device. There are currently a few caveats:

 * The OnePlus 6T bootloader is /bad/ and tries to select the OnePlus 6
   DTB.
 * There are two panel variants of the PocoPhone F1 which both have the
   same board-id.

Both of this issues will be solved by using a different bootloader like
u-boot which can select the correct DTB based on arbitrary matching, or
alternatively by fixing up the DTB in the kernel itself.