Fix root detection issue if the sd card also is luksed
On a setup where you got a fde on eMMC and also a luksed partition on
the sd card the find_root_partition can (and will on ppp) grab the wrong
one, making it impossible to boot the device.

To avoid this issue we will add a detection behavior, based on the BOOT_IMAGE
variable provided by tow-boot. It indicate which boot partition is used
and make it easy to check firstly the available root partition on the
same disk.