add pinephone-keyboard config
closes #1515

This patch HAS NOT been tested to see if it actually fixes #1515. I do
not have my pinephone keyboard with me and will not till v22.06 is out.
Can someone please test this patch before merging?

Note: this is being merged upstream but needs a refactor of the
Makefile. See:

https://lists.sr.ht/~mil/sxmo-devel/patches/32630

Since we cannot get the refactor of the Makefile out before the next
release, this is the next best thing.