main/sxmo-common: make sure sxmo-utils 1.11.1 always used with  superd 0.7
I added a patch to sxmo-utils 1.9.0 to work around bug [1] in superd. sxmo-utils
with the workaround for bug [1] was called Sxmo 1.9.1.
superd 0.7 fixes bug [1] where crashed service does not auto-restart if it
forked children [1,2].
Sxmo 1.11.1 removes the workaround for superd bug [1] in Sxmo.
Since Sxmo 1.11.1 is meant to be used with superd 0.7, add a hard
dependency for superd 0.7 and sxmo-utils in sxmo-common.

see:
[1] https://todo.sr.ht/~craftyguy/superd/38
[2] https://git.sr.ht/~craftyguy/superd/refs/0.7