main/sxmo-common: make sure sxmo-utils 1.9.1 always used with superd 0.6
I added a patch to sxmo-utils 1.9.0 to work around bug [1] in superd.
sxmo-utils with the workaround for bug [1] was called Sxmo 1.9.1 [3,4].
superd 0.7 fixes bug [1] where crashed service does not auto-restart if
it
forked children [1,2].
Since Sxmo 1.9.1 is meant to be used with superd 0.6, add a hard
dependency for superd 0.6 and sxmo-utils 1.9.1 in sxmo-common.

see:
[1] https://todo.sr.ht/~craftyguy/superd/38
[2] https://git.sr.ht/~craftyguy/superd/refs/0.7
[3] https://git.sr.ht/~mil/sxmo-utils/commit/350eaa09ce53e9225241f31b48b9ae2b08eaa8f6
[4] https://lists.sr.ht/~mil/sxmo-announce/%3C87bks1edaw.fsf%40momi.ca%3E