temp/libcamera: Add APKBUILD for libcamera v.0.0.2
Add APKBUILD for libcamera in temp/libcamera.

libcamera is built from the gitlab mirror:
https://gitlab.com/jmondi/libcamera
as long as it doesn't have a more official mirror.

Build libcamera with support for the simple pipeline handler (Pinephone)
and the RkISP1 pipeline handler (Pinephone Pro).

As commented in the APKBUILD, it is required not to strip binaries at
installation time to preserve the IPA module signature and avoid the IPA
process to be run isolated.

Signed-off-by: Jacopo Mondi <jacopo@jmondi.org>