community/purism-librem5: move from main
This device currently has 1 active maintainer (me), and therefore no
longer meets the ">=2 maintainers" requirement for being in the "Main"
device category. This is the only reason it is being moved, there is no
reduction in functionality with this change or any other reasons for the
move, implied or otherwise.

I hope that one day we can find a second person to help maintain it, so
we can move it back to "main."

[ci:skip-vercheck]