main/postmarketos-base-ui-networkmanager: use a second usb-tethering connection (MR 4400)
This is a follow-up of IRC discuss.

I'm investigating on USB Tethering mode for Sxmo, and how to help our users to
enable this mode.

Pmos rely on a NetworkManager dispatcher script to deal with tethering. Atm it
manage its owned system network manager connection. It check it ipv4.method
value, and toggle the tethering.

This is a proposition to use multiple system connections, so that a user could
switch from one to another mode without editing the connections themselves.