main/postmarketos-base: don't depend on alpine-base
We want to transition to systemd, and the prospects of alpine becoming
more flexible to accomodate for that are not very clear in the short-term.
So instead of waiting for all that[1], let's start preparing the process
and not depend on alpine-base. It should be no big deal, as that package
sees very little development anyway. Adding systemd alternative init will
be the job of the systemd branch.

[1] https://gitlab.alpinelinux.org/alpine/aports/-/merge_requests/63935